requirements.txt
streamlit
Pillow
gTTS
google-generativeai
python-dotenv
pytesseract